fixing double to float problem

+++ src/main/java/edu/uiuc/zenvisage/model/Point.java
@@ -1,16 +1,16 @@
 package edu.uiuc.zenvisage.model;
 
 public class Point {
-	double x;
-	double y;
+	float x;
+	float y;
 	
-	public double getX() {
+	public float getX() {
 		return x;
 	}
 	public void setX(int x) {
 		this.x = x;
 	}
-	public double getY() {
+	public float getY() {
 		return y;
 	}
 	public void setY(int y) {
